requirements.txt
# Core dependencies
pytest>=7.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
typing-extensions>=4.0.0

# TTS and Audio
gtts>=2.3.0
elevenlabs>=0.3.0

# Data validation
pydantic>=2.0.0

# AI and ML
google-generativeai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0

# Web framework
python-multipart>=0.0.6

# Authentication
bcrypt>=4.0.0
PyJWT>=2.8.0

# WebSocket client & test tools
websocket-client>=1.7.0
httpx>=0.27.0